% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmake.R
\name{groupmake}
\alias{groupmake}
\title{Assign people to groups while trying to consider their preferences}
\usage{
groupmake(data, target)
}
\arguments{
\item{data}{A matrix encoding group preferences, see \code{sim_data} for format}

\item{target}{An integer giving the exact amount of people each group should have}
}
\value{
A 0/1 matrix representing the final group assignments.
}
\description{
This function is guaranteed to assign people to groups of \code{target} size until
either people or available groups run out. If a group has more people who
want to be in it than the \code{target} size, a random selection will be made.
Some people may end up being assigned to groups they did not show a preference
for as a result and this result is not deterministic.
}
\examples{
library(grouputils)
set.seed(6)
data <- sim_data(8, 13, 0.2)
groupmake(data, 2)
}
