% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmake.R
\name{sim_data}
\alias{sim_data}
\title{Simulate a group preference matrix}
\usage{
sim_data(n_grp, n_ppl, p_choice = 0.5)
}
\arguments{
\item{n_grp}{Number of available gorups}

\item{n_ppl}{Number of people to sort into groups}

\item{p_choice}{Per-person probability of being interested in a given topic}
}
\value{
A 0/1 matrix encoding preferences. Groups as rows and people as columns.
}
\description{
Simulate a group preference matrix
}
\examples{
set.seed(6)
grouputils::sim_data(8, 13, 0.2)
}
